switch to QGram index for finding matches. move BitTree to it's own file, and fix up docs
