+++ src/data_structures/mod.rs
@@ -14,3 +14,4 @@ pub mod smallints;
 pub mod fmindex;
 pub mod qgram_index;
 pub mod interval_tree;
+pub mod bit_tree;
